Make command creation accessible by other tools
